requirements.txt
flask>=2.3.0
requests>=2.31.0
qrcode>=7.4.0
Pillow>=10.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
